// src/libs/enums/appErrors.ts
export enum AppErrors {
  INVALID_CHANNEL_URL = "INVALID_CHANNEL_URL",
  CHANNEL_NOT_FOUND = "CHANNEL_NOT_FOUND",
  ANALYSIS_FAILED = "ANALYSIS_FAILED",
  UNKNOWN_ERROR = "UNKNOWN_ERROR",
  CHANNEL_URL_REQUIRED = "CHANNEL_URL_REQUIRED",
  YOUTUBE_CHANNEL_REQUIRED = "YOUTUBE_CHANNEL_REQUIRED",
  INVALID_YOUTUBE_CHANNEL = "INVALID_YOUTUBE_CHANNEL",
  API_KEY_MISSING = "API_KEY_MISSING",
  API_REQUEST_FAILED = "API_REQUEST_FAILED",
  INVALID_API_RESPONSE = "INVALID_API_RESPONSE",
  VIDEOS_NOT_FOUND = "VIDEOS_NOT_FOUND",
}